Separate tun queue fd from interface accessors
This change supports applications that want to receive or pass a tun
device's file descriptor to another (possibly less privileged)
application and have it usable by tokio-tun on the other side.

The alternative is to open a multiqueue tun device and temporarily open
additional queues, which has a disruptive impact on existing flows
received on the tun device.